spectral/__init__.py
# Load environment variables
from pathlib import Path
from dotenv import load_dotenv

load_dotenv(
    dotenv_path=Path(__file__).resolve().parent / ".env"
)

from .polars_quant_expressions import *
from . import data
from . import charts

