pyproject.toml
[project]
name = "spectral"
version = "0.1.0"
description = "Quant Research Library"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bokeh>=3.8.1",
    "massive>=2.0.2",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "polars>=1.34.0",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
]
